refactor: moved newsletter funcionality to component in landing page, delete previous component
