+++ app/contact/page.tsx
@@ -1,5 +1,4 @@
 import ContactForm from "@/components/ui/contact-form";
-import NewsletterForm from "@/components/ui/newsletter-form";
 
 export const metadata = {
     title: "Superhero Store - Contact",
@@ -11,8 +10,6 @@ export default function ContactPage() {
     return (
         <>
             < ContactForm />
-
-            <NewsletterForm />
         </>
     )
 }
